Predict/requirements.txt
absl-py==0.9.0
aiohttp==3.4.4
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.0.4
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==18.2.0
backcall==0.1.0
Beaker==1.5.4
bleach==3.1.1
blessings==1.7
blivet==3.0.0b1
blivet-gui==2.1.9
botocore==1.15.36
bottle==0.12.13
Brlapi==0.6.7
cached-property==1.5.1
cachetools==4.0.0
CairoSVG==1.0.20
camelot-py==0.7.2
certifi==2019.6.16
cffi==1.11.5
chardet==3.0.4
chrome-printtopdf==0.0.2
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
codecov==2.0.15
colorama==0.3.9
coloredlogs==14.0
colour-runner==0.0.5
conan==1.9.2
conda==4.5.11
contextlib2==0.6.0.post1
cookies==2.2.1
coverage==4.5.3
coveralls==1.8.1
cryptography==2.3
cupshelpers==1.0
curses-ex==0.3
cycler==0.10.0
Cython==0.29.15
cytoolz==0.8.0
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.0.6
descartes==1.1.0
distro==1.1.0
Django==2.2
django-filter==2.1.0
django-split-settings==0.3.0
django-tastypie==0.14.2
djangorestframework==3.9.2
dnspython==1.16.0
docker==3.2.1
docker-compose==1.20.1
docker-pycreds==0.2.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
EasyProcess==0.2.5
Elixir==0.7.1
entrypoints==0.3
espatools==0.0.8
et-xmlfile==1.0.1
fasteners==0.14.1
Fiona==1.8.13.post1
Flask==1.0.2
Flask-WTF==0.14.2
fros==1.1
fsspec==0.7.1
funcsigs==1.0.2
future==0.16.0
gast==0.2.2
gcloud==0.17.0
geopandas==0.7.0
GOES==0.0.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.5.8
gpg==1.10.0
grpcio==1.26.0
gunicorn==19.9.0
h5py==2.10.0
httplib2==0.12.1
humanfriendly==7.1.1
humanize==0.5.1
idna==2.8
idna-ssl==1.1.0
imageio==2.5.0
imageio-ffmpeg==0.3.0
influxdb==5.2.2
iniparse==0.4
initial-setup==0.3.56
IPy==0.81
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isc==2.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.16.0
Jinja2==2.10
jmespath==0.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
jws==0.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
langtable==0.0.38
lazy-object-proxy==1.3.1
lxml==4.2.4
Mako==1.0.9.dev0
Markdown==3.1
markdown2==2.3.7
MarkupSafe==0.23
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==7.0.0
moviepy==1.0.0
multidict==4.4.2
munch==2.5.0
natsort==6.0.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
node-semver==0.2.0
nose==1.3.7
nose-htmloutput==0.6.0
notebook==6.0.3
notify2==0.3.1
ntplib==0.3.3
numpy==1.18.2
oauth2client==3.0.0
oauthlib==3.1.0
object-detection==0.0.3
olefile==0.46
opencv-python==4.0.0.21
openpyxl==2.6.2
opt-einsum==3.1.0
ordered-set==2.0.2
osmnx==0.11.4
packaging==18.0
paho-mqtt==1.4.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.1
patch==1.16
pathlib2==2.3.4
pbr==5.1.2
pdfminer.six==20181108
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
pika==1.1.0
Pillow==6.1.0
pluggy==0.9.0
pluginbase==0.7
ply==3.9
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==3.0.3
properties==0.5.6
protobuf==3.11.3
ptyprocess==0.6.0
PVGeo==2.0.1
pwquality==1.4.0
py==1.8.0
py2exe==0.9.2.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycairo==1.16.3
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycryptodome==3.4.3
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
pygame==1.9.4
Pygments==2.4.2
pygobject==3.28.3
pyinotify==0.9.6
PyIscsi==1.0
PyJWT==1.6.4
pykickstart==3.12
pylint==2.1.1
PyMuPDF==1.14.12
pyOpenSSL==17.3.0
pyparsing==2.4.0
pyparted==3.11.0
PyPDF2==1.26.0
pyproj==2.6.0
PyQt5==5.10.1
Pyrebase==3.0.27
PySocks==1.6.8
pystray==0.14.3
pytest==4.4.0
python-augeas==0.5.0
python-dateutil==2.8.0
python-dmidecode==3.12.2
python-etcd==0.4.5
python-jwt==2.0.1
python-meh==0.46
python-mimeparse==1.6.0
python-xlib==0.20
pytoml==0.1.18
pytz==2019.1
pyudev==0.21.0
PyVirtualDisplay==0.2.1
pyvista==0.21.0
pyxdg==0.25
PyYAML==5.1
pyzmq==18.1.1
qtconsole==4.6.0
registries==0.1
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
responses==0.9.0
rpm==4.14.2.1
rsa==4.0
Rtree==0.9.4
ruamel.yaml==0.13.14
s3fs==0.4.2
scipy==1.4.1
scooby==0.4.1
selenium==3.141.0
Send2Trash==1.5.0
sepolicy==1.1
setools==4.1.1
setroubleshoot==1.1
sh==1.12.14
Shapely==1.7.0
simpleline==1.1
sip==4.19.8
six==1.12.0
slip==0.6.4
slip.dbus==0.6.4
smartcols==0.3.0
sortedcontainers==2.1.0
sos==3.7
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
SSSDConfig==1.16.4
stringify==0.1.1
systemd-python==234
tb-nightly==1.14.0a20190603
Tempita==0.5.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-cpu==2.1.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.4.0
tf-estimator-nightly==1.14.0.dev2019060501
toolz==0.9.0
tornado==6.0.3
tqdm==4.28.1
traitlets==4.3.3
typed-ast==1.1.0
urllib3==1.25.3
utils==1.0.1
vectormath==0.2.2
virtualenv==16.0.0
virtualenv-api==2.1.16
vtk==8.1.2
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.47.0
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
xlrd==0.7.1
xlwt==0.7.2
xmltodict==0.12.0
yarl==1.2.6
